+++ config/packages/vich_uploader.yaml
@@ -1,6 +1,7 @@
 vich_uploader:
     db_driver: orm
+    storage: gaufrette
     mappings:
         backups:
-            uri_prefix: /files/backups
-            upload_destination: '%kernel.project_dir%/public/files/backups'
+            uri_prefix: '%uploads_base_url%'
+            upload_destination: backup_fs
