+++ tests/Command/SwitchStorageModeCommandTest.php
@@ -8,16 +8,17 @@ use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Console\Tester\CommandTester;
 
-final class SwitchToAwsStorageCommandTest extends KernelTestCase
+final class SwitchStorageModeCommandTest extends KernelTestCase
 {
-    public function testExecute(): void
+    public function testSwitchToAws(): void
     {
         $kernel = self::createKernel();
         $application = new Application($kernel);
 
-        $command = $application->find('app:switch-aws-storage');
+        $command = $application->find('app:switch-storage-mode');
         $commandTester = new CommandTester($command);
         $commandTester->setInputs([
+            'AWS S3',
             'aws_s3 access id',
             'aws_s3 access secret',
             'aws_s3 bucket name',
@@ -27,5 +28,21 @@ final class SwitchToAwsStorageCommandTest extends KernelTestCase
         $commandTester->execute(['command' => $command->getName()]);
         $output = $commandTester->getDisplay();
         self::assertStringContainsString('Successfully switched to AWS S3 storage!', $output);
+    }
+
+    public function testSwitchToLocal(): void
+    {
+        $kernel = self::createKernel();
+        $application = new Application($kernel);
+
+        $command = $application->find('app:switch-storage-mode');
+        $commandTester = new CommandTester($command);
+        $commandTester->setInputs([
+            'Locally',
+        ]);
+
+        $commandTester->execute(['command' => $command->getName()]);
+        $output = $commandTester->getDisplay();
+        self::assertStringContainsString('Successfully switched to Locally storage!', $output);
     }
 }
