--- tests/Command/SwitchToLocalStorageCommandTest.php
@@ -1,25 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Command;
-
-use Symfony\Bundle\FrameworkBundle\Console\Application;
-use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
-use Symfony\Component\Console\Tester\CommandTester;
-
-final class SwitchToLocalStorageCommandTest extends KernelTestCase
-{
-    public function testExecute(): void
-    {
-        $kernel = self::createKernel();
-        $application = new Application($kernel);
-
-        $command = $application->find('app:switch-local-storage');
-        $commandTester = new CommandTester($command);
-
-        $commandTester->execute(['command' => $command->getName()]);
-        $output = $commandTester->getDisplay();
-        self::assertStringContainsString('Successfully switched to local storage!', $output);
-    }
-}
